requirements.txt
torch==2.2.2
torchvision==0.17.2
streamlit==1.34.0
pillow==10.3.0
numpy==1.26.4